Remove dispatch from ChartLayoutContextProvider
