+++ scripts/snapshots/typesFiles.txt
@@ -90,6 +90,7 @@
   "types/state/RechartsStoreProvider.d.ts",
   "types/state/ReportBar.d.ts",
   "types/state/ReportChartProps.d.ts",
+  "types/state/ReportMainChartProps.d.ts",
   "types/state/ReportPolarOptions.d.ts",
   "types/state/SetGraphicalItem.d.ts",
   "types/state/SetLegendPayload.d.ts",
